Moving code around and renaming
* Renames FunctionLifter to MCToIRLifter
* Moves LiftCodeIntoModule() to a new Lift.(h|cpp)
* Renames old Lift.(h|cpp) to LegacyLift.(h|cpp)
* Removes LegacyLift.cpp from build
